extract out serve
